refactor(subconscious): consolidate heartbeat + scratchpad modules, add spawn_subagent, remove prompt.rs
- Add `spawn_subagent` to subconscious agent tools so it can delegate
  to the researcher agent for deep web/external research
- Move scratchpad into module folder: scratchpad.rs + scratchpad_tools.rs
  → scratchpad/mod.rs + scratchpad/tools.rs
- Move heartbeat module into subconscious/heartbeat/ (engine, rpc,
  schemas, planner). Old heartbeat/mod.rs becomes a thin re-export shim
- Remove prompt.rs — identity loading moved to engine.rs, archetype
  prompt now lives in agent/prompt.md
