+++ src/openhuman/subconscious/heartbeat/planner/collectors.rs
