+++ src/openhuman/subconscious/heartbeat/planner/utils.rs
